+++ apps/app/src/lib/components/CancelOrEndRecordingSessionButton.svelte
@@ -2,7 +2,7 @@
 	import WhisperingButton from '$lib/components/WhisperingButton.svelte';
 	import { Button } from '$lib/components/ui/button/index.js';
 	import * as Tooltip from '$lib/components/ui/tooltip/index.js';
-	import { mediaStreamManager } from '$lib/services/MediaRecorderService.svelte';
+	import { mediaStreamManager } from '$lib/services/MediaStreamService.svelte';
 	import { recorder } from '$lib/stores/recorder.svelte';
 	import { fasterRerecordExplainedDialog } from './FasterRerecordExplainedDialog.svelte';
