+++ apps/app/src/lib/services/MediaRecorderService.svelte.ts
@@ -1,182 +1,219 @@
-import { settings } from '$lib/stores/settings.svelte.js';
+import { mediaStreamManager } from '$lib/services/MediaStreamService.svelte';
+import { toast } from '$lib/services/ToastService';
+import { settings } from '$lib/stores/settings.svelte';
 import {
 	Ok,
-	WhisperingErr,
 	type WhisperingResult,
-	tryAsyncBubble,
+	WhisperingErr,
 	tryAsyncWhispering,
+	trySyncBubble,
+	trySyncWhispering,
+	type WhisperingRecordingState,
 } from '@repo/shared';
 import { nanoid } from 'nanoid/non-secure';
-import { toast } from './ToastService.js';
+import { renderErrAsToast } from './renderErrorAsToast';
 
-type MediaStreamManager = {
-	readonly isStreamValid: boolean;
-	getExistingStream(): Promise<WhisperingResult<MediaStream | null>>;
-	refreshStream(): Promise<WhisperingResult<MediaStream>>;
-	destroy(): void;
+const TIMESLICE_MS = 1000;
+
+type MediaRecorderService = {
+	readonly recordingState: WhisperingRecordingState;
+	startRecording: () => Promise<WhisperingResult<undefined>>;
+	stopRecording: () => Promise<WhisperingResult<Blob>>;
+	cancelRecording: () => Promise<WhisperingResult<undefined>>;
 };
 
-export const mediaStreamManager = createMediaStreamManager();
+export const mediaRecorder = createMediaRecorder();
 
-function createMediaStreamManager(): MediaStreamManager {
-	let currentStream: MediaStream | null = null;
-	let isStreamValid = $state<boolean>(false);
-	const setStream = (stream: MediaStream | null) => {
-		currentStream = stream;
-		isStreamValid = stream?.active ?? false;
+export function createMediaRecorder(): MediaRecorderService {
+	let recordingState = $state<WhisperingRecordingState>('IDLE');
+	let mediaRecorder: MediaRecorder | null = null;
+	const recordedChunks: Blob[] = [];
+
+	const setMediaRecorder = (value: MediaRecorder | null) => {
+		mediaRecorder = value;
+		updateRecordingState(value);
+	};
+
+	const updateRecordingState = (recorder: MediaRecorder | null) => {
+		if (!recorder) {
+			recordingState = 'IDLE';
+			return;
+		}
+
+		switch (recorder.state) {
+			case 'recording':
+				recordingState = 'RECORDING';
+				break;
+			case 'paused':
+				recordingState = 'IDLE';
+				break;
+			case 'inactive':
+				recordingState = 'IDLE';
+				break;
+		}
+	};
+
+	const resetRecorder = () => {
+		recordedChunks.length = 0;
+		setMediaRecorder(null);
+		if (!settings.value.isFasterRerecordEnabled) {
+			mediaStreamManager.destroy();
+		}
 	};
 
 	return {
-		get isStreamValid() {
-			return isStreamValid;
+		get recordingState() {
+			return recordingState;
 		},
-		async getExistingStream(): Promise<WhisperingResult<MediaStream | null>> {
-			if (currentStream === null) return Ok(null);
-			if (!currentStream.active) {
-				toast.warning({
-					title: 'Open stream is inactive',
-					description: 'Refreshing recording session...',
-				});
-				return this.refreshStream();
+
+		async startRecording() {
+			if (mediaRecorder) {
+				return renderErrAsToast(
+					WhisperingErr({
+						title: 'Unexpected media recorder already exists',
+						description:
+							'It seems like it was not properly deinitialized after the previous stopRecording or cancelRecording call.',
+						action: { type: 'none' },
+					}),
+				);
 			}
-			return Ok(currentStream);
-		},
-		async refreshStream() {
-			this.destroy();
 			const toastId = nanoid();
-			toast.loading({
-				id: toastId,
-				title: 'Connecting to selected audio input device...',
-				description: 'Please allow access to your microphone if prompted.',
-			});
-			if (!settings.value.selectedAudioInputDeviceId) {
-				toast.loading({
-					id: toastId,
-					title: 'No device selected',
-					description: 'Defaulting to first available audio input device...',
+
+			const getNewStream = async () => {
+				const newStreamResult = await mediaStreamManager.refreshStream();
+				if (!newStreamResult.ok) return newStreamResult;
+				const newStream = newStreamResult.data;
+				return Ok(newStream);
+			};
+
+			const getReusedStream = async () => {
+				const existingStreamResult =
+					await mediaStreamManager.getExistingStream();
+				if (!existingStreamResult.ok) return existingStreamResult;
+				const existingStream = existingStreamResult.data;
+				if (!existingStream) {
+					toast.loading({
+						id: toastId,
+						title: 'Error initializing media recorder with preferred device',
+						description:
+							'Trying to find another available audio input device...',
+					});
+					return await getNewStream();
+				}
+				return Ok(existingStream);
+			};
+
+			const newStreamResult = settings.value.isFasterRerecordEnabled
+				? await getReusedStream()
+				: await getNewStream();
+			if (!newStreamResult.ok) return newStreamResult;
+			const newStream = newStreamResult.data;
+
+			const newMediaRecorderResult: WhisperingResult<MediaRecorder> =
+				trySyncWhispering({
+					try: () => {
+						const recorder = new MediaRecorder(newStream, {
+							bitsPerSecond: Number(settings.value.bitrateKbps) * 1000,
+						});
+						recorder.addEventListener('start', () =>
+							updateRecordingState(recorder),
+						);
+						recorder.addEventListener('stop', () =>
+							updateRecordingState(recorder),
+						);
+						recorder.addEventListener('pause', () =>
+							updateRecordingState(recorder),
+						);
+						recorder.addEventListener('resume', () =>
+							updateRecordingState(recorder),
+						);
+						recorder.addEventListener('dataavailable', (event: BlobEvent) => {
+							if (!event.data.size) return;
+							recordedChunks.push(event.data);
+						});
+
+						return recorder;
+					},
+					catch: (error) => ({
+						_tag: 'WhisperingError',
+						title: 'Error initializing media recorder with preferred device',
+						description: 'Please try again',
+						action: { type: 'more-details', error },
+					}),
 				});
-				const firstAvailableStreamResult = await getFirstAvailableStream();
-				if (!firstAvailableStreamResult.ok) return firstAvailableStreamResult;
-				const firstAvailableStream = firstAvailableStreamResult.data;
-				setStream(firstAvailableStream);
-				toast.info({
-					id: toastId,
-					title: 'Defaulted to first available audio input device',
-					description: 'You can select a specific device in the settings.',
-				});
-				return Ok(firstAvailableStream);
-			}
-			const maybeStream = await getStreamForDeviceId(
-				settings.value.selectedAudioInputDeviceId,
-			);
-			if (maybeStream !== null) {
-				setStream(maybeStream);
-				toast.success({
-					id: toastId,
-					title: 'Connected to selected audio input device',
-					description: 'Successfully connected to your microphone stream.',
-				});
-				return Ok(maybeStream);
-			}
-			toast.loading({
-				id: toastId,
-				title: 'Error connecting to selected audio input device',
-				description: 'Trying to find another available audio input device...',
-			});
-			const firstAvailableStreamResult = await getFirstAvailableStream();
-			if (!firstAvailableStreamResult.ok) return firstAvailableStreamResult;
-			const firstAvailableStream = firstAvailableStreamResult.data;
-			setStream(firstAvailableStream);
-			toast.info({
-				id: toastId,
-				title: 'Defaulted to first available audio input device',
-				description: 'You can select a specific device in the settings.',
-			});
-			return Ok(firstAvailableStream);
+
+			if (!newMediaRecorderResult.ok) return newMediaRecorderResult;
+
+			const newMediaRecorder = newMediaRecorderResult.data;
+			newMediaRecorder.start(TIMESLICE_MS);
+			setMediaRecorder(newMediaRecorder);
+			return Ok(undefined);
 		},
-		destroy() {
-			if (currentStream === null) return;
-			for (const track of currentStream.getTracks()) {
-				track.stop();
+		async stopRecording() {
+			const stopResult: WhisperingResult<Blob> = await tryAsyncWhispering({
+				try: () =>
+					new Promise<Blob>((resolve, reject) => {
+						if (!mediaRecorder) {
+							reject(new Error('No active media recorder'));
+							return;
+						}
+						mediaRecorder.addEventListener('stop', () => {
+							if (!mediaRecorder) {
+								reject(
+									new Error('Media recorder was nullified during stop event'),
+								);
+								return;
+							}
+							const audioBlob = new Blob(recordedChunks, {
+								type: mediaRecorder.mimeType,
+							});
+							resolve(audioBlob);
+						});
+						mediaRecorder.stop();
+					}),
+				catch: (error) => ({
+					_tag: 'WhisperingError',
+					title: 'Error stopping media recorder',
+					description: 'Please try again',
+					action: {
+						type: 'more-details',
+						error,
+					},
+				}),
+			});
+			resetRecorder();
+			return stopResult;
+		},
+		async cancelRecording() {
+			if (!mediaRecorder) return Ok(undefined);
+			const cancelResult: WhisperingResult<undefined> =
+				await tryAsyncWhispering({
+					try: () =>
+						new Promise<undefined>((resolve, reject) => {
+							if (!mediaRecorder) {
+								reject(new Error('No active media recorder'));
+								return;
+							}
+							mediaRecorder.addEventListener('stop', () => {
+								resetRecorder();
+								resolve(undefined);
+							});
+							mediaRecorder.stop();
+						}),
+					catch: (error) => ({
+						_tag: 'WhisperingError',
+						title: 'Error canceling media recorder',
+						description: 'Please try again',
+						action: {
+							type: 'more-details',
+							error,
+						},
+					}),
+				});
+			if (!settings.value.isFasterRerecordEnabled) {
+				mediaStreamManager.destroy();
 			}
-			currentStream = null;
+			return cancelResult;
 		},
 	};
 }
-
-export const enumerateRecordingDevices = (): Promise<
-	WhisperingResult<MediaDeviceInfo[]>
-> =>
-	tryAsyncWhispering({
-		try: async () => {
-			const allAudioDevicesStream = await navigator.mediaDevices.getUserMedia({
-				audio: true,
-			});
-			const devices = await navigator.mediaDevices.enumerateDevices();
-			for (const track of allAudioDevicesStream.getTracks()) {
-				track.stop();
-			}
-			const audioInputDevices = devices.filter(
-				(device) => device.kind === 'audioinput',
-			);
-			return audioInputDevices;
-		},
-		catch: (error) => ({
-			_tag: 'WhisperingError',
-			title: 'Error enumerating recording devices',
-			description:
-				'Please make sure you have given permission to access your audio devices',
-			action: {
-				type: 'more-details',
-				error,
-			},
-		}),
-	});
-
-const getStreamForDeviceId = async (recordingDeviceId: string) => {
-	const result = await tryAsyncBubble({
-		try: async () => {
-			const stream = await navigator.mediaDevices.getUserMedia({
-				audio: {
-					deviceId: { exact: recordingDeviceId },
-					channelCount: 1, // Mono audio is usually sufficient for voice recording
-					sampleRate: 16000, // 16 kHz is a good balance for voice
-				},
-			});
-			return stream;
-		},
-		catch: (error) => ({
-			_tag: 'GetStreamError',
-			message:
-				error instanceof Error ? error.message : 'Please try again later.',
-		}),
-	});
-	if (result.ok) return result.data;
-	return null;
-};
-
-const getFirstAvailableStream = async (): Promise<
-	WhisperingResult<MediaStream>
-> => {
-	const recordingDevicesResult = await enumerateRecordingDevices();
-	if (!recordingDevicesResult.ok) return recordingDevicesResult;
-	const recordingDevices = recordingDevicesResult.data;
-
-	for (const device of recordingDevices) {
-		const maybeStream = await getStreamForDeviceId(device.deviceId);
-		if (maybeStream) {
-			settings.value.selectedAudioInputDeviceId = device.deviceId;
-			mediaStreamManager.refreshStream();
-			return Ok(maybeStream);
-		}
-	}
-	return WhisperingErr({
-		title: 'No available audio input devices',
-		description: 'Please make sure you have a microphone connected',
-		action: {
-			type: 'link',
-			label: 'Open Settings',
-			goto: '/settings',
-		},
-	});
-};
